implemented on delete authors

+++ controllers/authors.js
@@ -51,14 +51,19 @@ const createNewAuthor = asyncFunction(async (req, res) => {
 //////////////////////////////////// delete author ///////////////////////////////////////
 
 const deleteAuthorById = asyncFunction(async (req, res) => {
-  const author = await Author.findByIdAndDelete({ _id: req.params.authorId });
-  // const books = await Book.findOne({ authorId : req.params.authorId}).deleteMany();
-  // console.log(books)
+  const author = await Author.findById({ _id: req.params.authorId });
   if (!author) {
     throw { status: 404, message: 'Author not found!' };
   }
-  res.status(200).send(`Deleted author: ${author.firstName}`);
-  // and his books
+  const authorBooks = await Book.find({ authorId: req.params.authorId })
+  console.log(authorBooks);
+  if(authorBooks.length!==0)
+  {
+    throw{status:409 , message :"You cannot delete an author without deleting his books first!"}
+  }
+  const author2 = await Author.findByIdAndDelete({ _id: req.params.authorId });
+  res.status(200).send(author2);
+ 
 });
 
 //////////////////////////////////// update author ///////////////////////////////////////
